requirements.txt
# Django and related packages
Django==4.1.5
djangorestframework==3.13.1

# Gunicorn for running the application
gunicorn==22.0.0

# PostgreSQL database adapter
psycopg2-binary==2.9.6

# drf_yasg for Swagger documentation
drf_yasg==1.20.0

# rest_framework_swagger for Swagger UI
file:django_rest_swagger-2.2.0-py2.py3-none-any.whl

# Additional common dependencies
django-cors-headers==3.13.0
django-environ==0.9.0
djangorestframework-simplejwt==5.2.2